requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
PyYAML==6.0.1
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
